added product data

+++ bs-service-app/src/main/java/com/condigence/service/CartService.java
@@ -25,5 +25,5 @@ public interface CartService {
 
 	void checkout() throws NotEnoughProductsInStockException;
 
-	BigDecimal getTotal();
+	Integer getTotal();
 }
